+++ ta4j-core/src/main/java/org/ta4j/core/analysis/criteria/WinningPositionsRatioCriterion.java
@@ -37,22 +37,13 @@ public class WinningPositionsRatioCriterion extends AbstractAnalysisCriterion {
 
     @Override
     public Num calculate(BarSeries series, Position position) {
-        return isProfitablePosition(series, position) ? series.numOf(1) : series.numOf(0);
-    }
-
-    private boolean isProfitablePosition(BarSeries series, Position position) {
-        if (position.isClosed()) {
-            Num zero = series.numOf(0);
-            return position.getProfit().isGreaterThan(zero);
-        }
-        return false;
+        return position.hasProfit() ? series.numOf(1) : series.numOf(0);
     }
 
     @Override
     public Num calculate(BarSeries series, TradingRecord tradingRecord) {
-        long numberOfProfitable = tradingRecord.getPositions().stream().filter(t -> isProfitablePosition(series, t))
-                .count();
-        return series.numOf(numberOfProfitable).dividedBy(series.numOf(tradingRecord.getPositionCount()));
+        long numberOfProfitPositions = tradingRecord.getPositions().stream().filter(Position::hasProfit).count();
+        return series.numOf(numberOfProfitPositions).dividedBy(series.numOf(tradingRecord.getPositionCount()));
     }
 
     @Override